requirements.txt
fastapi[standard]==0.124.0
pydantic==2.8.0
sqlmodel==0.0.22
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
python-dotenv==1.0.1
email-validator==2.2.0
